# modules/trading-mcp/contract/capital-gateway.openapi.json
{
  "components": {
    "schemas": {
      "Account": {
        "properties": {
          "active": {
            "default": false,
            "title": "Active",
            "type": "boolean"
          },
          "available": {
            "title": "Available",
            "type": "number"
          },
          "balance": {
            "title": "Balance",
            "type": "number"
          },
          "currency": {
            "title": "Currency",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "pnl": {
            "title": "Pnl",
            "type": "number"
          }
        },
        "required": [
          "id",
          "name",
          "currency",
          "balance",
          "available",
          "pnl"
        ],
        "title": "Account",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "Resolution": {
        "description": "Candle time frame. One vocabulary serves REST and streaming alike — the two take the\nsame spellings.",
        "enum": [
          "MINUTE",
          "MINUTE_5",
          "MINUTE_15",
          "MINUTE_30",
          "HOUR",
          "HOUR_4",
          "DAY",
          "WEEK"
        ],
        "title": "Resolution",
        "type": "string"
      },
      "SetActiveAccount": {
        "properties": {
          "account_id": {
            "description": "an id from GET /accounts",
            "title": "Account Id",
            "type": "string"
          }
        },
        "required": [
          "account_id"
        ],
        "title": "SetActiveAccount",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "capital.com gateway — trading, deep history and a live stream behind one contract. Demo environment only. Endpoints return neutral DTOs; provider quirks (session tokens, the instrument tree, asynchronous settlement) stay inside the adapter. The WebSocket at /ws/stream is not described by this schema — see the module README.",
    "title": "TradingCenter · capital-gateway",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/accounts/active": {
      "put": {
        "description": "Switch the active account. Positions and orders act on it afterwards.\n\n**Switching drops the quote stream.** capital.com ends the streaming session when the\nfinancial account changes, so anything collecting candles through `/ws/quotes` sees a\ndisconnect and reconnects on its own — a gap of seconds, in data nobody is watching at\nthe moment of the switch. Said here because this route answers success while the\nconsequence lands somewhere else entirely.",
        "operationId": "set_active_account_accounts_active_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetActiveAccount"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Account"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Set Active Account",
        "tags": [
          "accounts"
        ]
      }
    },
    "/accounts/top-up": {
      "post": {
        "description": "Move the demo account's balance, and answer with the account as it stands after.\n\nActs on the **active** account — there is no account id to pass, because capital.com\nadjusts the session's own account and a parameter here would promise a choice that\ndoes not exist. Switch first if the money belongs somewhere else.",
        "operationId": "top_up_accounts_top_up_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TopUp"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Account"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Top Up",
        "tags": [
          "accounts"
        ]
      }
    },
...
    "/instruments": {
      "get": {
        "description": "Every instrument, deduped. `truncated` is true when the bound cut the walk short.",
        "operationId": "instruments_instruments_get",
        "parameters": [
          {
            "description": "how much of the tree to walk; defaults higher when asset_class is set",
            "in": "query",
            "name": "max_nodes",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maximum": 5000,
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "how much of the tree to walk; defaults higher when asset_class is set",
              "title": "Max Nodes"
            }
          },
          {
            "description": "narrow to one class, e.g. 'CRYPTO'; see GET /asset-classes",
            "in": "query",
            "name": "asset_class",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "narrow to one class, e.g. 'CRYPTO'; see GET /asset-classes",
              "title": "Asset Class"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InstrumentPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Instruments",
        "tags": [
          "market-data"
        ]
      }
    },
    "/instruments/search": {
      "get": {
        "operationId": "search_instruments_instruments_search_get",
        "parameters": [
          {
            "description": "e.g. 'gold', 'apple', 'btc'",
            "in": "query",
            "name": "q",
            "required": true,
            "schema": {
              "description": "e.g. 'gold', 'apple', 'btc'",
              "title": "Q",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Instrument"
                  },
                  "title": "Response Search Instruments Instruments Search Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Search Instruments",
        "tags": [
          "market-data"
        ]
      }
    },
    "/instruments/{symbol}/candles": {
      "get": {
        "description": "One request's worth of candles. For more than 1000, use `/history`.",
        "operationId": "candles_instruments__symbol__candles_get",
        "parameters": [
          {
            "in": "path",
            "name": "symbol",
            "required": true,
            "schema": {
              "title": "Symbol",
              "type": "string"
            }
          },
          {
            "description": "candle time frame",
            "in": "query",
            "name": "resolution",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/Resolution",
              "default": "MINUTE",
              "description": "candle time frame"
            }
          },
          {
            "description": "the provider's ceiling is 1000",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "description": "the provider's ceiling is 1000",
              "maximum": 1000,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Candle"
                  },
                  "title": "Response Candles Instruments  Symbol  Candles Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Candles",
        "tags": [
          "market-data"
        ]
      }
    },
    "/instruments/{symbol}/history": {
      "get": {
        "description": "Candles paged past the provider's per-request ceiling.\n\nThis is a long request by design — 20 000 five-minute candles measured at 30 provider\ncalls and 26 seconds. The response says how many requests it took, and paging stops\nearly if the client disconnects.",
        "operationId": "history_instruments__symbol__history_get",
        "parameters": [
          {
            "in": "path",
            "name": "symbol",
            "required": true,
            "schema": {
              "title": "Symbol",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "resolution",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/Resolution",
              "default": "MINUTE_5"
            }
          },
          {
            "description": "how many candles to reach back for",
            "in": "query",
            "name": "bars",
            "required": false,
            "schema": {
              "default": 1000,
              "description": "how many candles to reach back for",
              "maximum": 50000,
              "minimum": 1,
              "title": "Bars",
              "type": "integer"
            }
          },
          {
            "description": "reach back from this instant instead of now, so a window that ended in the past can be requested directly",
            "in": "query",
            "name": "before",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "reach back from this instant instead of now, so a window that ended in the past can be requested directly",
              "title": "Before"
            }
          },
          {
            "description": "stop here: no candle older than this is fetched or returned. `bars` counts candles, so for an instrument that is not open around the clock it spans more calendar time than it looks — this is the only way to name a lower bound in time rather than in candles",
            "in": "query",
            "name": "after",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "stop here: no candle older than this is fetched or returned. `bars` counts candles, so for an instrument that is not open around the clock it spans more calendar time than it looks — this is the only way to name a lower bound in time rather than in candles",
              "title": "After"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CandleHistory"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "History",
        "tags": [
          "market-data"
        ]
      }
    },
    "/instruments/{symbol}/terms": {
      "get": {
        "description": "The deposit and size rules for one instrument — no price; that is `/search` and\n`/candles`. A field the provider omits comes back null rather than as a default.",
        "operationId": "instrument_terms_instruments__symbol__terms_get",
        "parameters": [
          {
            "in": "path",
            "name": "symbol",
            "required": true,
            "schema": {
              "title": "Symbol",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InstrumentTerms"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Instrument Terms",
        "tags": [
          "market-data"
        ]
      }
    },
    "/orders": {
      "post": {
        "description": "Place an order. MARKET fills now (FILLED); LIMIT and STOP rest (WORKING).\n\nThe answer is settled, not acknowledged. A deal the provider has not resolved comes\nback PENDING with its reference — never FILLED.",
        "operationId": "place_order_orders_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlaceOrderRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Order"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Place Order",
        "tags": [
          "trading"
        ]
      }
    },
...
  }
}